backend/requirements.txt
Flask>=2.0.1
Pillow>=9.0.0
numpy>=1.21.2
requests>=2.26.0
python-dotenv>=0.19.0
Flask-SQLAlchemy>=2.5.1
Flask-Cors>=3.0.10
gunicorn>=20.1.0
opencv-python>=4.5.3.56
cryptography>=41.0.0
pytest>=7.0.0
python-magic>=0.4.27
marshmallow>=3.19.0
ratelimit>=2.2.1
logging-utils>=0.3.0